requirements369.txt
absl-py==0.15.0
asn1crypto==0.24.0
astunparse==1.6.3
attrs==22.1.0
bayesian-optimization==1.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.9.14
charset-normalizer==2.0.12
clang==5.0
cryptography==2.1.4
dataclasses==0.8
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==2.6
importlib-metadata==4.8.3
iniconfig==1.1.1
joblib==1.1.1
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
libclang==14.0.6
Markdown==3.3.7
numpy==1.19.5
oauthlib==3.2.1
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.5
psutil==5.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
PyGObject==3.26.1
pyparsing==3.0.9
pytest==7.0.1
python-dateutil==2.8.2
pytz==2024.1
pyxdg==0.25
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.2
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
threadpoolctl==3.1.0
tomli==1.2.3
torch==1.10.1
torchvision==0.11.2
typing-extensions==3.7.4.3
urllib3==1.26.12
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
